+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.3.2] - 2022-09-08
+
+### Fixed
+
+- Force query parsing and rebuild for URI to prevent malformed URI query
+
 ## [2.3.1] - 2022-06-21
 
 ### Fixed